refactor(test_ai): moved to separate files
